requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
Django==5.1.6
django-debug-toolbar==5.0.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-stubs==3.15.3
filelock==3.17.0
identify==2.6.8
idna==3.10
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.2
